MediumMovieInfoModel built
Since JSON parsing will easily get 6 fields at the same time (they are
in the same one JSONArray), it will cost more in simply building a
SimpleMovieInfoModel to parse only two fields and drop other 4. Field
movieImageUrl is currently considered as a String not a String array,
just to make it easier (only the main poster will be considered), which
will be left for improvement in the future.