Database Management completed
Every detail of a movie will be stored in the database, but TrailerUrl
and MovieReviews will only be parsed and stored when that movie gets
clicked. Before any movie gets clicked, a MediumMovieInfoModel will be
parsed and used for GridView display on the starting screen. Once
clicked, the movie info will be completed and a CompleteMovieInfoModel
will be applied to keep data in MovieDetailActivity.